requirements.txt
        Python 3.7 
        PyTorch
        pandas
        Flask
        streamlit
